# apps/backend/.openapi.after.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthetic Data Platform",
    "description": "Synthetic Data Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/sources/": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Upload Source",
        "description": "Upload and parse a DDL or schema JSON file.\n\n- **project_id**: UUID of the project (from path)\n- **file**: DDL SQL file (.sql) or schema JSON file (.json)\n- **dialect**: SQL dialect (postgres/mysql/mssql/hana) - only used for DDL files\n\nReturns parsed schema with warnings.",
        "operationId": "upload_source_api_v1_projects__project_id__sources__post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_api_v1_projects__project_id__sources__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/sources/{source_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Source Schema",
        "description": "Get parsed schema for a source.\n\nReturns normalized schema JSON and DAG.",
        "operationId": "get_source_schema_api_v1_projects__project_id__sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/sources/{source_id}/dag": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Source Dag",
        "description": "Get dependency DAG for a source schema.\n\nReturns the table dependency graph.",
        "operationId": "get_source_dag_api_v1_projects__project_id__sources__source_id__dag_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/sources/{source_id}/tables": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Source Tables",
        "description": "Get list of tables in a source schema.",
        "operationId": "get_source_tables_api_v1_projects__project_id__sources__source_id__tables_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableSchema"
                  },
                  "title": "Response Get Source Tables Api V1 Projects  Project Id  Sources  Source Id  Tables Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_source_api_v1_projects__project_id__sources__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "dialect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialect",
            "default": "postgres"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_source_api_v1_projects__project_id__sources__post"
      },
      "CanonicalSchema": {
        "properties": {
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableSchema"
            },
            "type": "array",
            "title": "Tables",
            "description": "Table definitions"
          },
          "dag": {
            "$ref": "#/components/schemas/DAGSchema",
            "description": "Table dependency graph"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Parsing warnings"
          }
        },
        "type": "object",
        "title": "CanonicalSchema",
        "description": "Canonical schema representation parsed from DDL or JSON."
      },
...
      "CheckConstraint": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Constraint name"
          },
          "expression": {
            "type": "string",
            "title": "Expression",
            "description": "Check constraint expression"
          }
        },
        "type": "object",
        "required": [
          "expression"
        ],
        "title": "CheckConstraint",
        "description": "Check constraint definition."
      },
...
      "ColumnSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Column name"
          },
          "dtype": {
            "type": "string",
            "title": "Dtype",
            "description": "Data type (normalized)"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable",
            "description": "Whether column allows NULL",
            "default": true
          },
          "pii_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pii Tag",
            "description": "PII classification tag"
          }
        },
        "type": "object",
        "required": [
          "name",
          "dtype"
        ],
        "title": "ColumnSchema",
        "description": "Column definition in a table schema."
      },
      "DAGSchema": {
        "properties": {
          "nodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nodes",
            "description": "Table names"
          },
          "edges": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Edges",
            "description": "Foreign key edges [from_table, to_table]"
          }
        },
        "type": "object",
        "title": "DAGSchema",
        "description": "Directed Acyclic Graph representation of table dependencies."
      },
...
      "ForeignKey": {
        "properties": {
          "from_col": {
            "type": "string",
            "title": "From Col",
            "description": "Source column name"
          },
          "to_table": {
            "type": "string",
            "title": "To Table",
            "description": "Target table name"
          },
          "to_col": {
            "type": "string",
            "title": "To Col",
            "description": "Target column name"
          }
        },
        "type": "object",
        "required": [
          "from_col",
          "to_table",
          "to_col"
        ],
        "title": "ForeignKey",
        "description": "Foreign key constraint definition."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SchemaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Schema ID"
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id",
            "description": "Associated source ID"
          },
          "schema": {
            "$ref": "#/components/schemas/CanonicalSchema",
            "description": "Parsed schema definition"
          },
          "dag": {
            "$ref": "#/components/schemas/DAGSchema",
            "description": "Table dependency graph"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Parsing warnings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Schema creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "schema",
          "dag",
          "created_at"
        ],
        "title": "SchemaResponse",
        "description": "API response for schema retrieval.\n\nNote: The field name \"schema\" conflicts with a BaseModel attribute in Pydantic.\nTo avoid shadowing warnings while preserving the public API shape, we use an\ninternal name and expose \"schema\" via field aliases."
      },
      "SourceUploadResponse": {
        "properties": {
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id",
            "description": "Created source ID"
          },
          "schema_id": {
            "type": "string",
            "format": "uuid",
            "title": "Schema Id",
            "description": "Created schema ID"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Source kind (ddl/json)"
          },
          "tables_count": {
            "type": "integer",
            "title": "Tables Count",
            "description": "Number of tables parsed"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Parsing warnings"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "schema_id",
          "kind",
          "tables_count"
        ],
        "title": "SourceUploadResponse",
        "description": "Response after uploading and parsing a source file."
      },
      "TableSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Table name"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnSchema"
            },
            "type": "array",
            "title": "Columns",
            "description": "Column definitions"
          },
          "pk": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pk",
            "description": "Primary key column names"
          },
          "uniques": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Uniques",
            "description": "Unique constraint column lists"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckConstraint"
            },
            "type": "array",
            "title": "Checks",
            "description": "Check constraints"
          },
          "fks": {
            "items": {
              "$ref": "#/components/schemas/ForeignKey"
            },
            "type": "array",
            "title": "Fks",
            "description": "Foreign key constraints"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TableSchema",
        "description": "Table definition with columns and constraints."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}